issue #3: merge in one bot
