+++ Dockerfile
@@ -9,9 +9,8 @@ RUN apt-get update && \
     apt-get upgrade -y && \
     apt-get install -y ffmpeg
 
-RUN apt-get --fix-missing
 # Make port 80 available to the world outside this container
 EXPOSE 80
 
 # Install any needed packages specified in requirements.txt
-CMD pip install --no-cache-dir -r requirements.txt && python app/main.py
+CMD pip install --no-cache-dir -r requirements.txt && python app/skinwalker_bot.py
