+++ settings.json
@@ -1,1 +1,1 @@
-{"appearance_frequency": 99}
+{"appearance_frequency": 30}
